# swagger/v2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API V2",
    "version": "v2"
  },
  "paths": {
    "/api/v2/vehicles": {
      "post": {
        "summary": "Creates a vehicle",
        "tags": [
          "Vehicles"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "business_associate_id": "vehicle",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate_id": {
                  "type": "integer"
                },
                "plate_number": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "seats": {
                  "type": "integer"
                },
                "ac": {
                  "type": "string"
                },
                "fuel_type": {
                  "type": "string"
                },
                "colour": {
                  "type": "string"
                },
                "fitness_doc_url": {
                  "type": "string"
                }
              },
              "required": [
                "business_associate_id",
                "plate_number",
                "model"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Vehicle created"
          },
          "422": {
            "description": "invalid request"
          }
        }
      },
      "get": {
        "summary": "Retrieves vehicles",
        "tags": [
          "Vehicles"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "Vehicles found",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate_id": {
                  "type": "integer"
                },
                "plate_number": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "seats": {
                  "type": "integer"
                },
                "ac": {
                  "type": "string"
                },
                "fuel_type": {
                  "type": "string"
                },
                "colour": {
                  "type": "string"
                },
                "fitness_doc_url": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "business_associate_id",
                "licence_number"
              ]
            }
          },
          "404": {
            "description": "Vehicle not found"
          }
        }
      }
    },
    "/api/v2/drivers": {
      "post": {
        "summary": "Creates a driver",
        "tags": [
          "Drivers"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "business_associate": "driver",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate": {
                  "type": "integer"
                },
                "licence_number": {
                  "type": "string"
                },
                "f_name": {
                  "type": "string"
                },
                "aadhaar_mobile_number": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string"
                },
                "marital_status": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "blood_group": {
                  "type": "string"
                }
              },
              "required": [
                "licence_number",
                "f_name",
                "aadhaar_mobile_number",
                "business_associate",
                "aadhaar_mobile_number",
                "date_of_birth",
                "marital_status",
                "gender",
                "blood_group"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Driver created"
          },
          "422": {
            "description": "invalid request"
          }
        }
      },
      "get": {
        "summary": "Retrieves drivers",
        "tags": [
          "Drivers"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "Drivers found",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate": {
                  "type": "integer"
                },
                "licence_number": {
                  "type": "string"
                },
                "f_name": {
                  "type": "string"
                },
                "aadhaar_mobile_number": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string"
                },
                "marital_status": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "blood_group": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "business_associate",
                "licence_number"
              ]
            }
          },
          "404": {
            "description": "driver not found"
          }
        }
      }
    },
    "/api/v2/drivers/{id}": {
      "get": {
        "summary": "Retrieves a driver",
        "tags": [
          "Drivers"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Driver found",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate": {
                  "type": "integer"
                },
                "licence_number": {
                  "type": "string"
                },
                "f_name": {
                  "type": "string"
                },
                "aadhaar_mobile_number": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string"
                },
                "marital_status": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "blood_group": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "business_associate",
                "licence_number"
              ]
            }
          },
          "404": {
            "description": "driver not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a driver",
        "tags": [
          "Drivers"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Driver deleted",
            "schema": {
              "type": "object",
              "properties": {
                "id": "id",
                "business_associate": {
                  "type": "integer"
                },
                "licence_number": {
                  "type": "string"
                },
                "f_name": {
                  "type": "string"
                },
                "aadhaar_mobile_number": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string"
                },
                "marital_status": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "blood_group": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "business_associate",
                "licence_number"
              ]
            }
          },
          "404": {
            "description": "driver not found"
          }
        }
      }
    },
    "/api/v2/drivers/search": {
      "get": {
        "summary": "Search driver or vehicle",
        "tags": [
          "Search driver or vehicle"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "type": "string",
            "in": "type"
          },
          {
            "name": "licence_number",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Search result",
            "schema": {
              "type": "object",
              "properties": {
                "id": "id",
                "business_associate": {
                  "type": "integer"
                },
                "licence_number": {
                  "type": "string"
                },
                "f_name": {
                  "type": "string"
                },
                "aadhaar_mobile_number": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string"
                },
                "marital_status": {
                  "type": "string"
                },
                "gender": {
                  "type": "string"
                },
                "blood_group": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "f_name",
                "licence_number"
              ]
            }
          },
          "404": {
            "description": "Record not found"
          }
        }
      }
    },
    "/api/v2/vehicles/{id}": {
      "delete": {
        "summary": "Delete a vehicle",
        "tags": [
          "Vehicles"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle deleted",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate_id": {
                  "type": "integer"
                },
                "plate_number": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "seats": {
                  "type": "integer"
                },
                "ac": {
                  "type": "string"
                },
                "fuel_type": {
                  "type": "string"
                },
                "colour": {
                  "type": "string"
                },
                "fitness_doc_url": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "business_associate_id",
                "plate_number"
              ]
            }
          },
          "404": {
            "description": "Vehicle not found"
          }
        }
      },
      "get": {
        "summary": "Retrieves a vehicle",
        "tags": [
          "Vehicles"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle found",
            "schema": {
              "type": "object",
              "properties": {
                "business_associate_id": {
                  "type": "integer"
                },
                "plate_number": {
                  "type": "string"
                },
                "model": {
                  "type": "string"
                },
                "seats": {
                  "type": "integer"
                },
                "ac": {
                  "type": "string"
                },
                "fuel_type": {
                  "type": "string"
                },
                "colour": {
                  "type": "string"
                },
                "fitness_doc_url": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "business_associate_id",
                "plate_number"
              ]
            }
          },
          "404": {
            "description": "Vehicle not found"
          }
        }
      }
    }
  }
}